+++ src/RcppExports.cpp
@@ -2347,14 +2347,15 @@ BEGIN_RCPP
 END_RCPP
 }
 // semi_implicit_integration
-void semi_implicit_integration(List network, double dt, NumericVector opt);
-RcppExport SEXP _medfate_semi_implicit_integration(SEXP networkSEXP, SEXP dtSEXP, SEXP optSEXP) {
+void semi_implicit_integration(List network, double dt, NumericVector opt, String cavitationRefill);
+RcppExport SEXP _medfate_semi_implicit_integration(SEXP networkSEXP, SEXP dtSEXP, SEXP optSEXP, SEXP cavitationRefillSEXP) {
 BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< List >::type network(networkSEXP);
     Rcpp::traits::input_parameter< double >::type dt(dtSEXP);
     Rcpp::traits::input_parameter< NumericVector >::type opt(optSEXP);
-    semi_implicit_integration(network, dt, opt);
+    Rcpp::traits::input_parameter< String >::type cavitationRefill(cavitationRefillSEXP);
+    semi_implicit_integration(network, dt, opt, cavitationRefill);
     return R_NilValue;
 END_RCPP
 }
@@ -5094,7 +5095,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_medfate_invincgam", (DL_FUNC) &_medfate_invincgam, 3},
     {"_medfate_gammds", (DL_FUNC) &_medfate_gammds, 2},
     {"_medfate_initCochardNetworks", (DL_FUNC) &_medfate_initCochardNetworks, 1},
-    {"_medfate_semi_implicit_integration", (DL_FUNC) &_medfate_semi_implicit_integration, 3},
+    {"_medfate_semi_implicit_integration", (DL_FUNC) &_medfate_semi_implicit_integration, 4},
     {"_medfate_initSperryNetworks", (DL_FUNC) &_medfate_initSperryNetworks, 1},
     {"_medfate_profitMaximization", (DL_FUNC) &_medfate_profitMaximization, 4},
     {"_medfate_parcohort", (DL_FUNC) &_medfate_parcohort, 5},